Print prompt when receive SIGINT
Note that in this commit we do this manually: capturing SIGINT,
and print a newline then the prompt. This may not be the correct
solution. In bash, ^C is echoed but ^Z is not. `stty` says echoctl
flag is on, meaning that echoing is enabled. So later we may have
to tackle this problem.