Change $jobs to an array
This enables us to maintain a list such that the newly
stopped job can be put at the head of the list, and `fg`
without arguments will bring the first job in the list
to the foreground.

This also makes it possible that `fg` and (later) `kill`
commands can specify a specific job with the number, which
is actually the index in the list.