Job continue and fix
Now jobs can be continued with `fg` builtin command. The behavior
is like in Bash, where the current and the previous jobs are
recorded.

Job waiting path is fixed. Shell needs to wait for a job because
that job is put to the foreground. Thus we call wait right after
a certain job has been put to the foreground, namely, after being
run and continued.