Fix SIGCHLD handling
We set up a handler for SIGCHLD, which increments a counter
tracking the number of SIGCHLD signals we received. When waiting
for a job, we only wait for the foreground processes. These
children could be:

1. terminated normally
2. terminated by a signal
3. stopped by a signal

In the first two cases, we should receive SIGCHLD, we reap the
children and decrement the counter. In the last case, we continue
waiting for other children until all alive children in the job are
stopped. Finally we return the terminal to the shell.