Fix reaping stopped then terminated children
Background children may be first stopped and then terminated,
like

    $ cat &         # stopped by SIGTTIN
    $ jobs          # get PID of cat, say 1234
    $ kill -9 1234  # reaped

But what we're doing is like this:

    if (all children are stopped)
        ...
    elsif (all children are terminated)
        ...
    end

In the `cat` case the child is killed when it's stopped. So
with this logic we only regard it as stopped, althoug it's
already reaped. So it won't be removed from the job list,
causing troubles later.